updating logs message

+++ src/database/tools.py
@@ -12,12 +12,8 @@ Self = TypeVar("Self", bound="Table")
 
 def commit(func):
     def wrapped(*args, **kwargs):
-        try:
-            func(*args, **kwargs)
-            session.commit()
-        except Exception as e:
-            print(f'[LOG] Failed while running {func.__name__}.')
-            print(f'\t---> {e}')
+        func(*args, **kwargs)
+        session.commit()
 
     return wrapped
 
@@ -26,7 +22,10 @@ class Table:
     @classmethod
     @commit
     def create(cls, *objects) -> None:
-        t = f'[DATABASE] Adding {len(objects)} object(s) ' \
+        if not objects:
+            return
+
+        t = f'Adding {len(objects)} object(s) ' \
             f'into {cls.__tablename__} table...'
         for obj in alive_it(objects, title=t):
             new_object = (cls(**obj)
@@ -50,7 +49,7 @@ class Table:
         objects = cls.read(where=where)
 
         if not objects:
-            print('[LOG] Tried to update a non-existent'
+            print('Tried to update a non-existent'
                   f'row from {cls.__tablename__}')
             return
 
@@ -63,7 +62,12 @@ class Table:
     def delete(cls, *, where: dict) -> None:
         objects = cls.read(where=where)
 
-        t = f'[DATABASE] Deleting {len(objects)} object(s) ' \
+        if not objects:
+            print('Tried to delete a non-existent'
+                  f'row from {cls.__tablename__}')
+            return
+
+        t = f'Deleting {len(objects)} object(s) ' \
             f'from {cls.__tablename__} table...'
         for obj in alive_it(objects, title=t):
             session.delete(obj)
